extract Method for RV in ProfileFragment
